fix: improve convctl copy logic for different node_modules paths

+++ convctl.sh
@@ -2,6 +2,10 @@
 
 [ -d .processes ] || mkdir .processes
 if [ ! -f ./convctl ]; then
-  cp ./node_modules/@corezoid/convctl/bin/convctl ./convctl
+  if [ -f ./node_modules/@corezoid/convctl/bin/convctl ]; then
+    cp ./node_modules/@corezoid/convctl/bin/convctl ./convctl
+  elif [ -f ../node_modules/@corezoid/convctl/bin/convctl ]; then
+    cp ../node_modules/@corezoid/convctl/bin/convctl ./convctl
+  fi
 fi
 ./convctl "$@"
